+++ src/components/typeahead-suggestion/TypeaheadSuggestion.test.ts
@@ -1,4 +1,4 @@
-import { mount } from '@vue/test-utils';
+import { mount, shallowMount } from '@vue/test-utils';
 import WvuiTypeaheadSuggestion from './TypeaheadSuggestion.vue';
 import { SearchResult } from '../typeahead-search/http/SearchClient';
 import * as suggestionsList from './TypeaheadSuggestion.stories.json';
@@ -18,24 +18,37 @@ describe( 'matches the snapshot', () => {
 	} );
 } );
 
-it( 'should focus/blur on active/inactive states', () => {
+it( 'should focus/blur on active/inactive states', async () => {
 	const div = document.createElement( 'div' );
+
 	div.id = 'root';
 	document.body.appendChild( div );
+
 	const wrapper = mount( WvuiTypeaheadSuggestion, {
-		attachTo: '#root'
+		attachTo: '#root',
+		propsData: {
+			suggestion: suggestionsList.pages[ 1 ]
+		}
 	} );
 	const el = wrapper.element as HTMLElement;
 
-	console.log(el);
-
-	wrapper.setData( { active: true } );
+	await wrapper.setProps( { active: true } );
 
 	expect( el ).toBe( document.activeElement );
 
-	wrapper.destroy()
+	await wrapper.setProps( { active: false } );
 
-	// wrapper.setData( { active: false } );
-	//
-	// expect( el ).not.toBe( document.activeElement );
+	expect( el ).not.toBe( document.activeElement );
+
+} );
+
+it( 'emits mouseover event', () => {
+	const wrapper = shallowMount( WvuiTypeaheadSuggestion, {
+		propsData: {
+			suggestion: suggestionsList.pages[ 1 ]
+		}
+	} );
+
+	wrapper.get( '.wvui-typeahead-suggestion' ).trigger( 'mouseover' );
+	expect( wrapper.emitted().mouseover ).toBeTruthy();
 } );